fix broken imports after services folder restructure
Relative imports in moved service files still pointed one level too
shallow after the providers/observations/infrastructure split — e.g.
../utils/ resolved to src/services/utils/ instead of src/utils/, and
./redis.service.mjs no longer existed in the providers/ directory.
