+++ src/services/providers/smhi.service.mjs
@@ -1,6 +1,6 @@
-import { SMHI_WPT_API_URL, SMHI_FORECAST_API_URL } from "../utils/constants.mjs";
-import { withCache } from "./redis.service.mjs";
-import userAgent from "../utils/userAgent.mjs";
+import { SMHI_WPT_API_URL, SMHI_FORECAST_API_URL } from "../../utils/constants.mjs";
+import { withCache } from "../infrastructure/redis.service.mjs";
+import userAgent from "../../utils/userAgent.mjs";
 
 const SMHI_FORECAST_CACHE_TTL = 600; // 10 minutes
